Update comments for Modifier.edgeSwipeToDismiss (#724)
* Update comments around use of Modifier.edgeSwipeToDismiss

Modifier.edgeSwipeToDismiss must exist within a SwipeToDismissBox that uses the same SwipeToDismissBoxState. Otherwise, edgeSwipeToDismiss waits indefinitely for the anchors to be provided that would tell it the swipeable limits.

* Update comments around use of Modifier.edgeSwipeToDismiss

Modifier.edgeSwipeToDismiss must exist within a SwipeToDismissBox that uses the same SwipeToDismissBoxState. Otherwise, edgeSwipeToDismiss waits indefinitely for the anchors to be provided that would tell it the swipeable limits.

Co-authored-by: stevebower <stevebower@google.com>