Removed scrolling delay from Rotary algorithm.
Scrolling delay is used for filtering negative values at the end of RSB scroll. It was done for a proper fling tracking.
This delay was also causing scrolling to have a pretty long delay when direction of the scroll was reversed.

This delay was removed and was left only for Fling tracking and not for scrolling