Integrate ExoPlayer / Media3-ui playback suppression on unsuitable output solution.
The playback suppression solution on unsuitable output helps handling the cases of preventing unintended playback on the builtin speaker of watch.
This includes launching the system output switcher when playback is suppression due to this reasom, resuming the playback automatically after the situation of unavailability of the suitable outputs is resolved, and also pausing the playback automatically when all the suitable outputs have been removed during an ongoing playback.