Modify VolumeUiState.Unknown to avoid rendering crashes.
VolumUiState.Unknown was added in https://github.com/google/horologist/pull/2512 without modifying the code reading it. RotaryVolumeControls has since been updated, but it still means that if it's used with any composable but RotaryVolumeControls it will crash when rendering those composables.

Since the value is already introduced and is used this PR changes the object to avoid these crashes. Changing Unknown to current=0, max=1, min=0 avoids the crashes, but in the unusual event that an app would offer this odd binary volume control it would mean that it would be interpreted as Unknown. To support this corner case, as well as the concept of Unknown this PR also adds a private boolean, unknown, that's only set by the Unknown instance. The unusual .also{} mechanism is to avoid exposing this boolean in the constructor.

This PR avoids any possible crashes if VolumeUiState is used incorrectly, but the composables using VolumeUiState should most likely be updated to handle Unknown too, now that it is introduced. Possibly VolumeUiState should be changed to a sealed interface with actual volume and Unknown as separate sub-classes, but that's a way larger change and out of scope for this de-risking PR.